+++ src/components/messages/Messages.tsx
@@ -17,7 +17,7 @@ import {
   useCallback,
 } from "react";
 import { io } from "socket.io-client";
-import { IUser } from "../../interfaces/IUser";
+import { OnlineUser } from "../../interfaces/OnlineUser";
 import { useNavigate } from "react-router-dom";
 import { useSelector } from "react-redux";
 import { Message, ReduxState, Rooms } from "../../redux/interfaces";
@@ -28,8 +28,9 @@ import Convo, { DeleteConversations } from "./Conversation";
 import useAuthGuard from "../../lib/index";
 import API from "../../lib/API";
 import OnlineUsers from "./OnlineUsers";
-import "./styles.scss";
 import { StartConversation } from "./StartConversation";
+import "./styles.scss";
+
 
 const ioAddress = process.env.REACT_APP_IO_URL!;
 
@@ -46,7 +47,7 @@ const Messages = () => {
   const [media, setMedia] = useState<string>("");
   const [isTyping, setIsTyping] = useState(false);
 
-  const [onlineUsers, setOnlineUsers] = useState<IUser[]>([]);
+  const [onlineUsers, setOnlineUsers] = useState<OnlineUser[]>([]);
   const [message, setMessage] = useState("");
   const [currentChat, setCurrentChat] = useState<Rooms | null>(null);
 
@@ -57,7 +58,7 @@ const Messages = () => {
   const [selectedIndex, setSelectedIndex] = useState<number>();
   const scrollRef = createRef<HTMLDivElement>();
 
-  const [openConvo, setOpenConvo] = useState(false)
+  const [openConvo, setOpenConvo] = useState(false);
 
   const socket = useMemo(() => {
     return io(ioAddress, { transports: ["websocket"] });
@@ -111,7 +112,7 @@ const Messages = () => {
       console.log("Connection established!");
     });
 
-    socket.on("getUsers", (users: IUser[]) => {
+    socket.on("getUsers", (users: OnlineUser[]) => {
       setOnlineUsers(users);
     });
