+++ src/components/messages/OnlineUsers.tsx
@@ -1,4 +1,4 @@
-import { IUser } from "../../interfaces/IUser";
+import { OnlineUser } from "../../interfaces/OnlineUser";
 import { Rooms, User } from "../../redux/interfaces";
 import { defaultAvatar } from "../../assets/icons";
 import { Image } from "react-bootstrap";
@@ -6,7 +6,7 @@ import API from "../../lib/API";
 import { Socket } from "socket.io-client";
 
 interface OnlineUsersProps {
-  onlineUsers: IUser[];
+  onlineUsers: OnlineUser[];
   currentUser: User;
   conversation: Rooms[];
   currentChat: Rooms | null;
@@ -26,7 +26,7 @@ export default function OnlineUsers({
   socket,
   setOpenConvo,
 }: OnlineUsersProps) {
-  const handleClick = async (friend: IUser) => {
+  const handleClick = async (friend: OnlineUser) => {
     if (!currentChat?.members.includes(currentUser)) {
       // here we should check if there's already a conversation going with the selected guy
       const conversationAlreadyOpened = conversation
@@ -54,7 +54,7 @@ export default function OnlineUsers({
     }
   };
 
-  const newConversation = async (friend: IUser) => {
+  const newConversation = async (friend: OnlineUser) => {
     try {
       const { data } = await API.post("/rooms", {
         senderId: currentUser._id,
