--- src/interfaces/IMessage.ts
@@ -1,11 +1,0 @@
-interface IMessage {
-  text: string
-  sender: string | undefined
-  receiver: string | undefined
-  image: string
-  media: string
-  socketId: string
-  timestamp: number
-}
-
-export default IMessage
