--- src/interfaces/IUser.ts
@@ -1,6 +1,0 @@
-export interface IUser {
-  _id: string
-  userName: string
-  image: string
-  socketId: string
-}
