+++ src/interfaces/OnlineUser.ts
@@ -1,0 +1,6 @@
+export interface OnlineUser {
+  _id: string;
+  userName: string;
+  image: string;
+  socketId: string;
+}
