+++ src/interfaces/WeatherData.ts
